Allow editing a contact.
This extracts a ContactForm component from the AddContact component.
It is shared by AddContact, for creating a new contact, and Contact,
for editing an existing contact.
ContactForm ensures that the form field ids are unique when editing multiple
contacts, by suffixing them with the id of the contact, or 'undefined' for the
add contact form.
Also remove AddContact's startsOpen property, which was only used for tests;
instead have the test click the 'Add contact' button to show the form when
needed.